# rubocop: reduce redundant block.call for performance